Update tippecanoe with proper zoom
This PR updates the tippecanoe script using the following two parameters, to resolve the issue with very sparse points rendering in the map;

`-zg or --maximum-zoom=g: Guess what is probably a reasonable maxzoom based on the spacing of features.`
`-B zoom or --base-zoom=zoom: Base zoom, the level at and above which all points are included in the tiles (default maxzoom). If you use -Bg, it will guess a zoom level that will keep at most 50,000 features in the densest tile. You can also specify a marker-width with -Bgwidth to allow fewer features in the densest tile to compensate for the larger marker, or -Bfnumber to allow at most number features in the densest tile.`

https://github.com/mapbox/tippecanoe
https://docs.mapbox.com/help/troubleshooting/large-data-tippecanoe/